+++ packages/sdk-core/tasks/testSchemasAndQueries.sh
@@ -6,9 +6,9 @@ JQ="npx --package=node-jq -- jq"
 set -xe
 
 if [ "$SUBGRAPH_RELEASE_TAG" == "feature" ];then
-    # we only support matic and goerli feature endpoints
+    # we only support matic and mumbai feature endpoints
     # however, we don't want to be blocked by matic for feature
-    NETWORKS=("goerli")
+    NETWORKS=("mumbai")
 fi
 
 if [ "$SUBGRAPH_RELEASE_TAG" == "dev" ] || [ "$SUBGRAPH_RELEASE_TAG" == "v1" ];then
@@ -31,7 +31,6 @@ for i in "${NETWORKS[@]}";do
     declare -A LEGACY_NETWORK_NAMES=(
         ["xdai-mainnet"]="xdai"
         ["polygon-mainnet"]="matic"
-        ["eth-goerli"]="goerli"
         ["polygon-mumbai"]="mumbai"
     )