+++ packages/metadata/CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [WIP]
 ### Added
 - toga and batchLiquidator for scroll-mainnet
+### Changed
+- removed eth-goerli and Görli based networks: optimism-goerli, arbitrun-goerli, base-goerli, polygon-zkevm-testnet
+### Fixed
+- Removed subgraph hosted endpoint entry for scroll-sepolia which doesn't exist
 
 ## [v1.1.27]
 ### Added